+++ crates/zng-wgt-layer/src/lib.rs
@@ -24,7 +24,9 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::{fmt, mem, ops};
 
+pub mod dialog;
 pub mod popup;
+
 struct LayersCtx {
     items: EditableUiNodeListRef,
 }